+++ src/FeatureToggleBundle/Exception/AlreadyDefinedException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace BestIt\FeatureToggleBundle\Exception;
+
+/**
+ * Class AlreadyDefinedException
+ *
+ * @author Michel Chowanski <chowanski@bestit-online.de>
+ * @package BestIt\FeatureToggleBundle\Exception
+ */
+class AlreadyDefinedException extends FeatureToggleException
+{
+}