+++ tests/FeatureToggleBundle/DependencyInjection/ConfigurationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BestIt\FeatureToggleBundle\Tests\DependencyInjection;
+namespace Tests\BestIt\FeatureToggleBundle\DependencyInjection;
 
 use BestIt\FeatureToggleBundle\DependencyInjection\Configuration;
 use PHPUnit\Framework\TestCase;
@@ -10,7 +10,7 @@ use Symfony\Component\Config\Definition\Builder\TreeBuilder;
  * Class ConfigurationTest
  *
  * @author Michel Chowanski <chowanski@bestit-online.de>
- * @package BestIt\FeatureToggleBundle\Tests\DependencyInjection
+ * @package Tests\BestIt\FeatureToggleBundle\DependencyInjection
  */
 class ConfigurationTest extends TestCase
 {